llm-editor-chat/backend/requirements.txt
fastapi==0.115.12
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.54
langchain_openai==0.3.14
openai==1.75.0
pydantic==2.11.3
pydantic-settings>=2.11,<3.0
python-dotenv==1.1.0
python_docx==1.1.2
python-multipart==0.0.6
fastapi-csrf-protect==0.3.2
